requirements.txt
google-genai>=1.0.0
pydantic
pandas
yfinance
numpy
streamlit
python-dotenv